+++ src/main/java/org/folio/circulation/resources/ExpiredSessionProcessingResource.java
@@ -9,7 +9,7 @@ import org.folio.circulation.infrastructure.storage.ConfigurationRepository;
 import org.folio.circulation.domain.notice.session.ExpiredSession;
 import org.folio.circulation.domain.notice.session.PatronActionSessionService;
 import org.folio.circulation.domain.notice.session.PatronActionType;
-import org.folio.circulation.infrastructure.storage.PatronExpiredSessionRepository;
+import org.folio.circulation.infrastructure.storage.sessions.PatronExpiredSessionRepository;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.ClockManager;
 import org.folio.circulation.support.Result;
